requirements.txt
ansible==5.6.0
diagrams==0.21.1
python-dotenv===0.20.0
argparse==1.4.0